+++ apps/www/package.json
@@ -70,7 +70,6 @@
     "babel-eslint": "10.1.0",
     "babel-jest": "26.0.1",
     "cross-env": "7.0.2",
-    "danger": "10.2.0",
     "eslint-config-react-app": "5.2.1",
     "eslint-plugin-css-modules": "2.11.0",
     "eslint-plugin-flowtype": "5.1.3",
