+++ package.json
@@ -32,6 +32,7 @@
     "@typescript-eslint/parser": "3.2.0",
     "all-contributors-cli": "6.16.0",
     "concurrently": "5.2.0",
+    "danger": "10.2.0",
     "eslint": "7.2.0",
     "eslint-config-prettier": "6.11.0",
     "eslint-plugin-import": "2.21.1",
