+++ tsconfig.json
@@ -22,7 +22,7 @@
     "**/*.spec.ts",
     "**/*.spec.tsx",
     "**/*.test.ts",
-    "**/*.test.tsx"
-    // "dangerfile.ts"
+    "**/*.test.tsx",
+    "dangerfile.ts"
   ]
 }
